ServiceImplementations.CRUD was lacking some love and will also now throw NotFoundExceptions.

+++ Tests/Standard.AspNetCore.TestApplication/Standard.AspNetCore.TestApplication.Application/Implementation/InvoicesService.cs
@@ -7,6 +7,7 @@ using AutoMapper;
 using Intent.RoslynWeaver.Attributes;
 using Standard.AspNetCore.TestApplication.Application.Interfaces;
 using Standard.AspNetCore.TestApplication.Application.Invoices;
+using Standard.AspNetCore.TestApplication.Domain.Common.Exceptions;
 using Standard.AspNetCore.TestApplication.Domain.Entities;
 using Standard.AspNetCore.TestApplication.Domain.Repositories;
 
@@ -36,35 +37,50 @@ namespace Standard.AspNetCore.TestApplication.Application.Implementation
                 Number = dto.Number,
             };
             _invoiceRepository.Add(newInvoice);
-            await _invoiceRepository.UnitOfWork.SaveChangesAsync();
+            await _invoiceRepository.UnitOfWork.SaveChangesAsync(cancellationToken);
             return newInvoice.Id;
         }
 
         [IntentManaged(Mode.Fully, Body = Mode.Fully)]
         public async Task<InvoiceDto> FindInvoiceById(Guid id, CancellationToken cancellationToken = default)
         {
-            var element = await _invoiceRepository.FindByIdAsync(id);
+            var element = await _invoiceRepository.FindByIdAsync(id, cancellationToken);
+
+            if (element is null)
+            {
+                throw new NotFoundException($"Could not find Invoice {id}");
+            }
             return element.MapToInvoiceDto(_mapper);
         }
 
         [IntentManaged(Mode.Fully, Body = Mode.Fully)]
         public async Task<List<InvoiceDto>> FindInvoices(CancellationToken cancellationToken = default)
         {
-            var elements = await _invoiceRepository.FindAllAsync();
+            var elements = await _invoiceRepository.FindAllAsync(cancellationToken);
             return elements.MapToInvoiceDtoList(_mapper);
         }
 
         [IntentManaged(Mode.Fully, Body = Mode.Fully)]
         public async Task UpdateInvoice(Guid id, InvoiceUpdateDto dto, CancellationToken cancellationToken = default)
         {
-            var existingInvoice = await _invoiceRepository.FindByIdAsync(id);
+            var existingInvoice = await _invoiceRepository.FindByIdAsync(id, cancellationToken);
+
+            if (existingInvoice is null)
+            {
+                throw new NotFoundException($"Could not find Invoice {id}");
+            }
             existingInvoice.Number = dto.Number;
         }
 
         [IntentManaged(Mode.Fully, Body = Mode.Fully)]
         public async Task DeleteInvoice(Guid id, CancellationToken cancellationToken = default)
         {
-            var existingInvoice = await _invoiceRepository.FindByIdAsync(id);
+            var existingInvoice = await _invoiceRepository.FindByIdAsync(id, cancellationToken);
+
+            if (existingInvoice is null)
+            {
+                throw new NotFoundException($"Could not find Invoice {id}");
+            }
             _invoiceRepository.Remove(existingInvoice);
         }
